+++ src/main/java/Coupons/Logic/UsersController.java
@@ -45,6 +45,16 @@ public class UsersController {
 	}
 	
 	public LoginData login(LoginForm loginForm) throws ApplicationException {
+		if (loginForm == null)
+			throw new ApplicationException(ErrorType.EMPTY, ErrorType.EMPTY.getInternalMessage(), false);
+
+		NameUtils.isValidName(loginForm.getUserName());
+		PasswordUtils.isValidPassword(loginForm.getPassword());
+
+		if (!usersDao.isUserNameExist(loginForm.getUserName())) {
+			throw new ApplicationException(ErrorType.USERNAME_DOES_NOT_EXISTS, ErrorType.USERNAME_DOES_NOT_EXISTS.getInternalMessage(),false);
+		}
+		
 		UserData userData = usersDao.login(loginForm.getUserName(),loginForm.getPassword());
 		int token = generateEncryptedToken(loginForm.getUserName());
 		cacheManager.put(token, userData);
@@ -60,49 +70,94 @@ public class UsersController {
 		return token.hashCode();
 	}
 
-	public long createUser(User user) throws ApplicationException {
+	public long createUser(User user, UserData userData) throws ApplicationException {
 		if (user == null) {
-			throw new ApplicationException(ErrorType.EMPTY, ErrorType.EMPTY.getInternalMessage());
+			throw new ApplicationException(ErrorType.EMPTY, ErrorType.EMPTY.getInternalMessage(), false);
 		}
+		
+		if (!userData.getType().name().equals("Administrator")) {
+			if (!user.getType().name().equals("Customer"))
+				throw new ApplicationException(ErrorType.USER_TYPE_MISMATCH, ErrorType.USER_TYPE_MISMATCH.getInternalMessage(), true);
+
+		}
+
 		
 		NameUtils.isValidName(user.getUserName());
 		PasswordUtils.isValidPassword(user.getPassword());
-
+		if (user.getId() < 1) {
+			throw new ApplicationException(ErrorType.INVALID_ID, ErrorType.INVALID_ID.getInternalMessage(), false);
+			
+		}
 		if (usersDao.getUserByMail(user.getEmail()) != null) {
-			throw new ApplicationException(ErrorType.EXISTING_EMAIL, ErrorType.EXISTING_EMAIL.getInternalMessage());
+			throw new ApplicationException(ErrorType.EXISTING_EMAIL, ErrorType.EXISTING_EMAIL.getInternalMessage(), false);
 		}
-		if ((user.getCompanyId() != 0 && user.getType().equals(ClientType.Customer))) {
-			throw new ApplicationException(ErrorType.COMPANY_ID_NOT_TYPE, ErrorType.COMPANY_ID_NOT_TYPE.getInternalMessage());
+		if ((user.getCompanyId() != null && user.getType().equals(ClientType.Customer))) {
+			throw new ApplicationException(ErrorType.COMPANY_ID_NOT_TYPE, ErrorType.COMPANY_ID_NOT_TYPE.getInternalMessage(), false);
 		}
-		if (user.getCompanyId() == 0 && user.getType().equals(ClientType.Company)) {
-			throw new ApplicationException(ErrorType.COMPANY_TYPE_NO_ID, ErrorType.COMPANY_TYPE_NO_ID.getInternalMessage());
+		if (user.getType().equals(ClientType.Company)) {
+			if(user.getCompanyId() == null) {
+			throw new ApplicationException(ErrorType.COMPANY_TYPE_NO_ID, ErrorType.COMPANY_TYPE_NO_ID.getInternalMessage(), false);
+			}
+			if (user.getCompanyId() < 1) {
+				throw new ApplicationException(ErrorType.INVALID_ID, ErrorType.INVALID_ID.getInternalMessage(), false);
+				
+			}
+			if(companiesDAO.isCompanyExists(user.getCompanyId())) {
+				throw new ApplicationException(ErrorType.COMPANY_ID_DOES_NOT_EXIST, ErrorType.COMPANY_ID_DOES_NOT_EXIST.getInternalMessage(), false);
+			}
 		}
+		
 		return usersDao.createUser(user);
 	}
 
 	
 
-	public void updateUser(User user) throws ApplicationException {
+	public void updateUser(User user, UserData userData) throws ApplicationException {
 
-		NameUtils.isValidName(user.getEmail());
+		if (user == null) {
+			throw new ApplicationException(ErrorType.EMPTY, ErrorType.EMPTY.getInternalMessage(), false);
+		}
+		
+		if (!userData.getType().name().equals("Administrator")) {
+			if (!user.getType().name().equals("Customer"))
+				throw new ApplicationException(ErrorType.USER_TYPE_MISMATCH, ErrorType.USER_TYPE_MISMATCH.getInternalMessage(), true);
+
+		}
+
+		
+		NameUtils.isValidName(user.getUserName());
 		PasswordUtils.isValidPassword(user.getPassword());
-
+		if (user.getId() < 1) {
+			throw new ApplicationException(ErrorType.INVALID_ID, ErrorType.INVALID_ID.getInternalMessage(), false);
+			
+		}
 		if (usersDao.getUserByID(user.getId()) == null) {
-			throw new ApplicationException(ErrorType.USER_ID_DOES_NOT_EXIST, ErrorType.USER_ID_DOES_NOT_EXIST.getInternalMessage());
+			throw new ApplicationException(ErrorType.USER_ID_DOES_NOT_EXIST, ErrorType.USER_ID_DOES_NOT_EXIST.getInternalMessage(), false);
+		}
+		if (usersDao.isUserNameExist(user.getUserName())) {
+			throw new ApplicationException(ErrorType.NAME_IS_ALREADY_EXISTS, ErrorType.NAME_IS_ALREADY_EXISTS.getInternalMessage(), false);
 		}
 		usersDao.updateUser(user);
 	}
 
-	public void deleteUser(long userId) throws ApplicationException {
+	public void deleteUser(long userId, UserData userData) throws ApplicationException {
+		if (!userData.getType().name().equals("Administrator")) {
+			if (userId != userData.getUserID())
+				throw new ApplicationException(ErrorType.USER_TYPE_MISMATCH, ErrorType.USER_TYPE_MISMATCH.getInternalMessage(), true);
 
-
+		}
+		if (userId < 1) {
+			throw new ApplicationException(ErrorType.INVALID_ID, ErrorType.INVALID_ID.getInternalMessage(), false);
+			
+		}
 		if (usersDao.getUserByID(userId) == null) {
-			throw new ApplicationException(ErrorType.USER_ID_DOES_NOT_EXIST, ErrorType.USER_ID_DOES_NOT_EXIST.getInternalMessage());
+			throw new ApplicationException(ErrorType.USER_ID_DOES_NOT_EXIST, ErrorType.USER_ID_DOES_NOT_EXIST.getInternalMessage(), false);
 		}
 		usersDao.deleteUserByID(userId); 
 	}
 
-	public void deleteUsersByCompanyId(long companyId) throws ApplicationException {
+	
+	/*public void deleteUsersByCompanyId(long companyId) throws ApplicationException {
 
 
 		if ( companiesDAO.getCompanyByID(companyId) == null) {
@@ -110,12 +165,35 @@ public class UsersController {
 		}
 		usersDao.deleteCompanysUsers(companyId);
 
-	}
+	}*/
 
-	public Collection<User> getAllUsers() throws ApplicationException
+	public Collection<User> getAllUsers(UserData userData) throws ApplicationException
 	{
+		if(!userData.getType().name().equals("Administrator"))
+		throw new ApplicationException(ErrorType.USER_TYPE_MISMATCH, ErrorType.USER_TYPE_MISMATCH.getInternalMessage(), true);
 
 		return usersDao.getAllUsers();
+
+	}
+	
+	
+	public User getUser(long userId, UserData userData) throws ApplicationException {
+
+		if (!userData.getType().name().equals("Administrator")) {
+			if (userId != userData.getUserID()) {
+				throw new ApplicationException(ErrorType.USER_TYPE_MISMATCH, ErrorType.USER_TYPE_MISMATCH.getInternalMessage(), true);
+			}
+		}
+		
+		if (userId < 1) {
+			throw new ApplicationException(ErrorType.INVALID_ID, ErrorType.INVALID_ID.getInternalMessage(), false);
+			
+		}
+		
+		if (!usersDao.isUserIDExist(userId)) {
+			throw new ApplicationException(ErrorType.USER_ID_DOES_NOT_EXIST, ErrorType.USER_ID_DOES_NOT_EXIST.getInternalMessage(),	false);
+		}
+		return usersDao.getUserByID(userId);
 
 	}
